test(pack): omit full runtime from declaration paths snapshot
