--- crates/pack-tests/tests/snapshot/typescript/dts_paths_fallback/assert.js
@@ -1,3 +1,0 @@
-const assert = require("node:assert/strict");
-
-assert.equal(require("./output/main.js").default, "dts-exact-runtime");
