+++ crates/pack-tests/tests/snapshot/typescript/dts_paths_fallback/output/19vvtx8fhyfb0.js
@@ -1,0 +1,26 @@
+(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([typeof document === "object" ? document.currentScript : undefined,
+"[project]/typescript/dts_paths_fallback/node_modules/dts-exact/runtime.js [client] (ecmascript)", ((__turbopack_context__) => {
+"use strict";
+
+function getMessage() {
+    return "dts-exact-runtime";
+}
+__turbopack_context__.s([
+    "f",
+    0,
+    getMessage
+]);
+}),
+"[project]/typescript/dts_paths_fallback/input/index.ts [client] (ecmascript)", ((__turbopack_context__) => {
+"use strict";
+
+var __TURBOPACK__imported__module__$5b$project$5d2f$typescript$2f$dts_paths_fallback$2f$node_modules$2f$dts$2d$exact$2f$runtime$2e$js__$5b$client$5d$__$28$ecmascript$29$__ = __turbopack_context__.i("[project]/typescript/dts_paths_fallback/node_modules/dts-exact/runtime.js [client] (ecmascript)");
+;
+var __TURBOPACK__default__export__ = (0, __TURBOPACK__imported__module__$5b$project$5d2f$typescript$2f$dts_paths_fallback$2f$node_modules$2f$dts$2d$exact$2f$runtime$2e$js__$5b$client$5d$__$28$ecmascript$29$__["f"])();
+__turbopack_context__.s([
+    "default",
+    0,
+    __TURBOPACK__default__export__
+]);
+}),
+]);
