+++ crates/pack-tests/tests/snapshot/typescript/dts_paths_fallback/output/turbopack-0_g3yyhy7khx8.js
@@ -1,0 +1,5 @@
+(globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([
+    typeof document === "object" ? document.currentScript : undefined,
+    {"otherChunks":["19vvtx8fhyfb0.js"],"runtimeModuleIds":["[project]/typescript/dts_paths_fallback/input/index.ts [client] (ecmascript)"]}
+]);
+// Dummy runtime
